Flattening fleet
