+++ fleet/coordinator/state_test.go
