+++ fleet/device/inventory.go
