+++ fleet/device/inventory_test.go
