+++ fleet/device/topology.go
